Moved Footer Assets
